Refactored ImageComponent into MediaComponent and ImageComponent.
The new VideoComponent will inherit from MediaComponent.
