+++ es-core/src/components/AnimatedImageComponent.cpp
@@ -20,7 +20,7 @@ void AnimatedImageComponent::load(const AnimationDef* def)
 		}
 
 		auto img = std::unique_ptr<ImageComponent>(new ImageComponent(mWindow));
-		img->setResize(mSize.x(), mSize.y());
+		img->setResize(mSize);
 		img->setImage(std::string(def->frames[i].path), false);
 		
 		mFrames.push_back(ImageFrame(std::move(img), def->frames[i].time));
@@ -43,7 +43,7 @@ void AnimatedImageComponent::onSizeChanged()
 {
 	for(auto it = mFrames.begin(); it != mFrames.end(); it++)
 	{
-		it->first->setResize(mSize.x(), mSize.y());
+		it->first->setResize(mSize);
 	}
 }
